refactor: consolidate suspicious repo validation into shared module
Extract duplicated suspicious owner names and route pairs from
src/services/github.ts and web/src/lib/repository-route-validation.ts
into a single shared module at src/core/suspicious-repo-patterns.ts.

Both the worker and web layers now import the same lists, eliminating
the risk of pattern drift between the two validation paths.

Closes #74
